MessageCard Components Update
